lib/mixer_module: split functions into separate headers
 - this arguably improves visibility/discoverability
