+++ src/lib/mixer_module/CMakeLists.txt
@@ -44,14 +44,25 @@ add_custom_command(OUTPUT ${functions_header}
 add_custom_target(output_functions_header DEPENDS ${functions_header})
 
 px4_add_library(mixer_module
+	functions/FunctionActuatorSet.hpp
+	functions/FunctionConstantMax.hpp
+	functions/FunctionConstantMin.hpp
+	functions/FunctionGimbal.hpp
+	functions/FunctionLandingGear.hpp
+	functions/FunctionManualRC.hpp
+	functions/FunctionMotors.hpp
+	functions/FunctionParachute.hpp
+	functions/FunctionServos.hpp
+
 	actuator_test.cpp
+	actuator_test.hpp
 	mixer_module.cpp
-	functions.cpp
+	mixer_module.hpp
 	)
 
 add_dependencies(mixer_module output_functions_header)
 target_compile_options(mixer_module PRIVATE ${MAX_CUSTOM_OPT_LEVEL})
-
+target_include_directories(mixer_module PRIVATE ${CMAKE_CURRENT_BINARY_DIR})
 
 px4_add_functional_gtest(SRC mixer_module_tests.cpp LINKLIBS mixer_module mixer)
